Django/website/requirements.txt
colorama==0.3.9
decorator==4.1.2
Django==1.11.3
django-cors-headers==2.1.0
djangorestframework==3.6.3
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
pickleshare==0.7.4
prompt-toolkit==1.0.15
Pygments==2.2.0
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
